requirements.txt
openai==1.3.5
flask==3.0.0
pandas==2.0.3
python-dotenv==1.0.0
langchain==0.0.340
mysql-connector-python==8.2.0
flask-cors==4.0.0
faiss-cpu==1.7.4
tiktoken==0.3.1
sklearn==0.0.post11
torch==2.1.1
transformers==4.35.2
pillow==10.1.0
gunicorn>=20.1.0